+++ Common/src/main/java/fuzs/easyanvils/core/CommonAbstractions.java
@@ -1,15 +1,14 @@
 package fuzs.easyanvils.core;
 
-import fuzs.easyanvils.handler.AnvilRepairContext;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AnvilMenu;
-
-import java.util.function.IntConsumer;
+import net.minecraft.world.item.ItemStack;
+import org.jetbrains.annotations.NotNull;
 
 public interface CommonAbstractions {
 
-    default AnvilRepairContext anvilRepairContextOf(AnvilMenu anvilMenu, Player player, String itemName, Container resultSlots, IntConsumer repairItemCountCost, IntConsumer cost) {
-        return new AnvilRepairContext(anvilMenu, player, itemName, resultSlots, repairItemCountCost, cost);
-    }
+    boolean onAnvilChange(AnvilMenu container, @NotNull ItemStack left, @NotNull ItemStack right, Container outputSlot, String name, int baseCost, Player player);
+
+    boolean isBookEnchantable(ItemStack inputStack, ItemStack bookStack);
 }