+++ Common/src/main/java/fuzs/easyanvils/world/inventory/ModAnvilMenu.java
@@ -1,23 +1,34 @@
 package fuzs.easyanvils.world.inventory;
 
+import fuzs.easyanvils.EasyAnvils;
+import fuzs.easyanvils.config.ServerConfig;
 import fuzs.easyanvils.core.ModServices;
-import fuzs.easyanvils.handler.AnvilRepairContext;
 import fuzs.easyanvils.init.ModRegistry;
 import fuzs.easyanvils.mixin.accessor.AnvilMenuAccessor;
 import fuzs.easyanvils.mixin.accessor.ItemCombinerMenuAccessor;
 import net.minecraft.Util;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.*;
+import net.minecraft.world.item.EnchantedBookItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.mutable.MutableInt;
+
+import java.util.Map;
+import java.util.function.IntConsumer;
 
 public class ModAnvilMenu extends AnvilMenu implements ContainerListener {
 
     public ModAnvilMenu(int id, Inventory inventory) {
-        // this is important, vanilla super adds a listener to the SimpleContainer which is required for the renaming edit box
-        // we have nothing in the other constructor that is required on the client anyways
+        // passing this to super instead of the other constructor is important, vanilla super adds a listener to the SimpleContainer which is required for the renaming edit box
+        // we have nothing in the other constructor that is required on the client anyway
         super(id, inventory);
     }
 
@@ -47,9 +58,193 @@ public class ModAnvilMenu extends AnvilMenu implements ContainerListener {
 
     @Override
     public void createResult() {
-        AnvilRepairContext context = ModServices.ABSTRACTIONS.anvilRepairContextOf(this, this.player, ((AnvilMenuAccessor) this).getItemName(), this.resultSlots, i -> ((AnvilMenuAccessor) this).setRepairItemCountCost(i), i -> this.setData(0, i));
-        if (!context.createResult(this.inputSlots.getItem(0), this.inputSlots.getItem(1))) {
-            super.createResult();
+        ItemStack left = this.inputSlots.getItem(0);
+        ItemStack right = this.inputSlots.getItem(1);
+        String itemName = ((AnvilMenuAccessor) this).getItemName();
+        MutableInt cost = new MutableInt();
+        this._createResult(left, right, itemName, cost, i -> ((AnvilMenuAccessor) this).setRepairItemCountCost(i));
+        this.setData(0, cost.intValue());
+    }
+
+    private void _createResult(ItemStack left, ItemStack right, String itemName, MutableInt cost, IntConsumer repairItemCountCost) {
+        cost.setValue(1);
+        if (left.isEmpty()) {
+            this.resultSlots.setItem(0, ItemStack.EMPTY);
+            cost.setValue(0);
+        } else {
+            ItemStack output = left.copy();
+            Map<Enchantment, Integer> map = EnchantmentHelper.getEnchantments(output);
+            int baseRepairCost = left.getBaseRepairCost() + (right.isEmpty() ? 0 : right.getBaseRepairCost());
+            // no prior work penalty, or fixed
+            baseRepairCost = EasyAnvils.CONFIG.get(ServerConfig.class).priorWorkPenalty.operator.applyAsInt(baseRepairCost);
+            repairItemCountCost.accept(0);
+            boolean isBook = false;
+
+            if (ModServices.ABSTRACTIONS.onAnvilChange(this, left, right, this.resultSlots, itemName, baseRepairCost, this.player)) {
+                return;
+            }
+
+            int repairOperationCost = 0;
+            int enchantOperationCost = 0;
+            int renameOperationCost = 0;
+            final int maxAnvilRepairCost = EasyAnvils.CONFIG.get(ServerConfig.class).maxAnvilRepairCost;
+            if (!right.isEmpty()) {
+                isBook = right.getItem() == Items.ENCHANTED_BOOK && !EnchantedBookItem.getEnchantments(right).isEmpty();
+                if (output.isDamageableItem() && output.getItem().isValidRepairItem(left, right)) {
+                    int l2 = Math.min(output.getDamageValue(), output.getMaxDamage() / 4);
+                    if (l2 <= 0) {
+                        this.resultSlots.setItem(0, ItemStack.EMPTY);
+                        cost.setValue(0);
+                        return;
+                    }
+
+                    int repairMaterials;
+                    for (repairMaterials = 0; l2 > 0 && repairMaterials < right.getCount(); ++repairMaterials) {
+                        int j3 = output.getDamageValue() - l2;
+                        output.setDamageValue(j3);
+                        ++repairOperationCost;
+                        l2 = Math.min(output.getDamageValue(), output.getMaxDamage() / 4);
+                    }
+
+                    repairItemCountCost.accept(repairMaterials);
+                } else {
+                    if (!isBook && (!output.is(right.getItem()) || !output.isDamageableItem())) {
+                        this.resultSlots.setItem(0, ItemStack.EMPTY);
+                        cost.setValue(0);
+                        return;
+                    }
+
+                    if (output.isDamageableItem() && !isBook) {
+                        int l = left.getMaxDamage() - left.getDamageValue();
+                        int i1 = right.getMaxDamage() - right.getDamageValue();
+                        int j1 = i1 + output.getMaxDamage() * 12 / 100;
+                        int k1 = l + j1;
+                        int l1 = output.getMaxDamage() - k1;
+                        if (l1 < 0) {
+                            l1 = 0;
+                        }
+
+                        if (l1 < output.getDamageValue()) {
+                            output.setDamageValue(l1);
+                            repairOperationCost += 2;
+                        }
+                    }
+
+                    Map<Enchantment, Integer> map1 = EnchantmentHelper.getEnchantments(right);
+                    boolean flag2 = false;
+                    boolean flag3 = false;
+
+                    for (Enchantment enchantment1 : map1.keySet()) {
+                        if (enchantment1 != null) {
+                            int otherEnchantmentLevel = map.getOrDefault(enchantment1, 0);
+                            int enchantmentLevel = map1.get(enchantment1);
+                            enchantmentLevel = otherEnchantmentLevel == enchantmentLevel ? enchantmentLevel + 1 : Math.max(enchantmentLevel, otherEnchantmentLevel);
+                            boolean compatibleWithItem = enchantment1.canEnchant(left);
+                            if (this.player.getAbilities().instabuild || left.is(Items.ENCHANTED_BOOK)) {
+                                compatibleWithItem = true;
+                            }
+
+                            for (Enchantment enchantment : map.keySet()) {
+                                if (enchantment != enchantment1 && !enchantment1.isCompatibleWith(enchantment)) {
+                                    compatibleWithItem = false;
+                                    ++enchantOperationCost;
+                                }
+                            }
+
+                            if (!compatibleWithItem) {
+                                flag3 = true;
+                            } else {
+                                flag2 = true;
+                                if (enchantmentLevel > enchantment1.getMaxLevel()) {
+                                    enchantmentLevel = enchantment1.getMaxLevel();
+                                }
+
+                                // prevent a level higher than max level from being lowered to max value
+                                if (EasyAnvils.CONFIG.get(ServerConfig.class).noAnvilMaxLevelLimit) {
+                                    int maxLevel = Math.max(map.getOrDefault(enchantment1, 0), map1.get(enchantment1));
+                                    maxLevel = Math.max(maxLevel, enchantmentLevel);
+                                    if (maxLevel != enchantmentLevel) {
+                                        enchantmentLevel = maxLevel;
+                                    }
+                                }
+
+                                map.put(enchantment1, enchantmentLevel);
+                                int rarityCostMultiplier = switch (enchantment1.getRarity()) {
+                                    case COMMON -> 1;
+                                    case UNCOMMON -> 2;
+                                    case RARE -> 4;
+                                    case VERY_RARE -> 8;
+                                };
+
+                                if (isBook) {
+                                    rarityCostMultiplier = Math.max(1, rarityCostMultiplier / 2);
+                                }
+
+                                enchantOperationCost += rarityCostMultiplier * enchantmentLevel;
+                                if (left.getCount() > 1) {
+                                    enchantOperationCost = maxAnvilRepairCost;
+                                }
+                            }
+                        }
+                    }
+
+                    if (flag3 && !flag2) {
+                        this.resultSlots.setItem(0, ItemStack.EMPTY);
+                        cost.setValue(0);
+                        return;
+                    }
+                }
+            }
+
+            if (StringUtils.isBlank(itemName)) {
+                if (left.hasCustomHoverName()) {
+                    renameOperationCost = EasyAnvils.CONFIG.get(ServerConfig.class).freeRenames.filter.test(left) ? 0 : 1;
+                    output.resetHoverName();
+                }
+            } else if (!itemName.equals(left.getHoverName().getString())) {
+                renameOperationCost = EasyAnvils.CONFIG.get(ServerConfig.class).freeRenames.filter.test(left) ? 0 : 1;
+                output.setHoverName(Component.literal(itemName));
+            }
+
+            if (isBook && !ModServices.ABSTRACTIONS.isBookEnchantable(output, right)) {
+                output = ItemStack.EMPTY;
+            }
+
+            cost.setValue(baseRepairCost + enchantOperationCost + repairOperationCost + renameOperationCost);
+
+            // when renaming is free make sure to let the item stack pass without being cleared
+            if (enchantOperationCost <= 0 && renameOperationCost == 0 && repairOperationCost == 0) {
+                output = ItemStack.EMPTY;
+            }
+
+            if (enchantOperationCost == 0 && cost.getValue() >= maxAnvilRepairCost && EasyAnvils.CONFIG.get(ServerConfig.class).alwaysRenameAndRepair) {
+                cost.setValue(maxAnvilRepairCost - 1);
+            }
+
+            // allow for custom max enchantment levels limit
+            if (cost.getValue() >= maxAnvilRepairCost && !this.player.getAbilities().instabuild) {
+                output = ItemStack.EMPTY;
+            }
+
+            if (!output.isEmpty()) {
+                int outputRepairCost = output.getBaseRepairCost();
+                if (!right.isEmpty() && outputRepairCost < right.getBaseRepairCost()) {
+                    outputRepairCost = right.getBaseRepairCost();
+                }
+
+                if (enchantOperationCost > 0 || !EasyAnvils.CONFIG.get(ServerConfig.class).alwaysRenameAndRepair) {
+                    outputRepairCost = AnvilMenu.calculateIncreasedRepairCost(outputRepairCost);
+                }
+
+                // don't add tag when there is no repair cost
+                if (outputRepairCost > 0) {
+                    output.setRepairCost(outputRepairCost);
+                }
+                EnchantmentHelper.setEnchantments(map, output);
+            }
+
+            this.resultSlots.setItem(0, output);
+            this.broadcastChanges();
         }
     }
 
@@ -83,5 +278,15 @@ public class ModAnvilMenu extends AnvilMenu implements ContainerListener {
     @Override
     public void dataChanged(AbstractContainerMenu containerMenu, int dataSlotIndex, int value) {
 
+    }
+
+    public static int repairCostToRepairs(int repairCost) {
+        repairCost++;
+        int repairs = 0;
+        while (repairCost >= 2) {
+            repairCost /= 2;
+            repairs++;
+        }
+        return repairs;
     }
 }